6.1

+++ pro-features/plus/meta-fields/db.php
@@ -4,7 +4,7 @@ defined('ABSPATH') or die("Jog on!");
 
 /**
  *
- * Return all meta fields for weight entry
+ * Return all meta enties for weight entry
  *
  * @param $entry_id
  * @return array
@@ -24,6 +24,24 @@ function ws_ls_meta( $entry_id ) {
 	$data = $wpdb->get_results( $sql, ARRAY_A );
 
     ws_ls_cache_user_set( 'meta-fields', $cache_key , $data, 30 );
+
+	return $data;
+}
+
+/**
+ *
+ * Return all meta entries for meta field
+ *
+ * @param $meta_field_id
+ * @return array
+ */
+function ws_ls_meta_for_given_meta_field( $meta_field_id ) {
+
+	global $wpdb;
+
+	$sql = $wpdb->prepare( 'Select * from ' . $wpdb->prefix . WE_LS_MYSQL_META_ENTRY . ' where meta_field_id = %d', $meta_field_id );
+
+	$data = $wpdb->get_results( $sql, ARRAY_A );
 
 	return $data;
 }
@@ -90,6 +108,26 @@ function ws_ls_meta_delete( $entry_id, $meta_field_id ) {
 
 	$result = $wpdb->delete( $wpdb->prefix . WE_LS_MYSQL_META_ENTRY, [ 'entry_id' => $entry_id, 'meta_field_id' => $meta_field_id ], [ '%d', '%d' ] );
 
+    do_action( 'wlt-meta-delete', $entry_id );
+
+	return ( 1 === $result );
+}
+
+/**
+ * Delete previous migrated values
+ *
+ * @return bool
+ */
+function ws_ls_meta_delete_migrated() {
+
+	if ( false === is_admin() ) {
+		return;
+	}
+
+	global $wpdb;
+
+	$result = $wpdb->delete( $wpdb->prefix . WE_LS_MYSQL_META_ENTRY, [ 'migrate' => 1 ], [ '%d' ] );
+
 	return ( 1 === $result );
 }
 
@@ -101,7 +139,13 @@ function ws_ls_meta_delete( $entry_id, $meta_field_id ) {
  */
 function ws_ls_meta_delete_for_entry( $entry_id ) {
 
+	if ( false === is_admin() ) {
+		return;
+	}
+
 	global $wpdb;
+
+    do_action( 'wlt-meta-entries-delete', $entry_id );
 
 	$result = $wpdb->delete( $wpdb->prefix . WE_LS_MYSQL_META_ENTRY, [ 'entry_id' => $entry_id ], [ '%d' ] );
 
@@ -116,6 +160,10 @@ function ws_ls_meta_delete_for_entry( $entry_id ) {
  * @return bool
  */
 function ws_ls_meta_delete_for_meta_field( $meta_field_id ) {
+
+	if ( false === is_admin() ) {
+		return;
+	}
 
 	global $wpdb;
 
@@ -193,6 +241,10 @@ function ws_ls_meta_fields( $exclude_system = true, $ignore_cache = false ) {
  */
 function ws_ls_meta_fields_update( $field ) {
 
+	if ( false === is_admin() ) {
+		return;
+	}
+
     // Ensure we have the expected fields.
     if ( false === ws_ls_meta_check_fields( $field, [ 'id', 'abv', 'field_name', 'field_type', 'suffix', 'mandatory', 'enabled' ] ) ) {
         return false;
@@ -216,6 +268,8 @@ function ws_ls_meta_fields_update( $field ) {
 
 	ws_ls_cache_user_delete( 'meta-fields' );
 
+	do_action( 'wlt-meta-fields-updating-meta-field', $id );
+
     if ( 1 === $result ) {
 
         // If the field type has changed in this update then delete existing data entries (as they won't relate to the new field type).
@@ -225,6 +279,8 @@ function ws_ls_meta_fields_update( $field ) {
 
         return true;
     }
+
+
 
     return false;
 }
@@ -243,6 +299,10 @@ function ws_ls_meta_fields_update( $field ) {
  * @return bool     true if success
  */
 function ws_ls_meta_fields_add( $field ) {
+
+	if ( false === is_admin() ) {
+		return;
+	}
 
     // Ensure we have the expected fields.
     if ( false === ws_ls_meta_check_fields( $field, [ 'abv', 'field_name', 'field_type', 'suffix', 'mandatory', 'enabled' ] ) ) {
@@ -273,7 +333,13 @@ function ws_ls_meta_fields_add( $field ) {
  */
 function ws_ls_meta_fields_delete( $id ) {
 
+	if ( false === is_admin() ) {
+		return;
+	}
+
     global $wpdb;
+
+    do_action( 'wlt-meta-fields-deleting-meta-field', $id );
 
     $result = $wpdb->delete( $wpdb->prefix . WE_LS_MYSQL_META_FIELDS, [ 'id' => $id ], [ '%d' ] );
 
@@ -333,6 +399,27 @@ function ws_ls_meta_fields_get_by_id( $id ) {
 }
 
 /**
+ * Fetch all user IDs that have a reference to this field (allows us to clear cache)
+ *
+ * @param $meta_field_id
+ */
+function ws_ls_meta_fields_get_user_ids_for_this_meta_field( $meta_field_id ) {
+
+    global $wpdb;
+
+    $sql = 'Select distinct weight_user_id from ' . $wpdb->prefix . WE_LS_MYSQL_META_ENTRY . ' e ' .
+        ' inner join ' . $wpdb->prefix . WE_LS_TABLENAME . ' d on e.entry_id = d.id 
+	         inner join ' . $wpdb->prefix . WE_LS_MYSQL_META_FIELDS . ' f on f.id = e.meta_field_id    
+	         where meta_field_id = %d';
+
+    $sql = $wpdb->prepare( $sql, $meta_field_id );
+
+    $results = $wpdb->get_results($sql, ARRAY_A );
+
+    return ( false === empty( $results ) ) ? wp_list_pluck( $results, 'weight_user_id' ) : $results;
+}
+
+/**
  * Return data formats
  *
  * @param $data
@@ -356,7 +443,8 @@ function ws_ls_meta_formats( $data ) {
 		'suffix' => '%s',
         'enabled' => '%d',
         'sort' => '%d',
-        'mandatory' => '%d'
+        'mandatory' => '%d',
+        'hide_from_shortcodes' => '%d'
     ];
 
     $return = [];
